Fix a potential crash.
When the delegate is set, we prompt the delegate for a `viewForAnnotation:`, which we use for measurement purposes. The annotation we pass is a dummy annotation, but the delegate doesn't know that, and it may ask the annotation for more details, perhaps by asking for the `originalAnnotations`. Since the delegate has no way of knowing that this is a dummy annotation (and should not have to care), we should avoid crashing when the try to do the "right" thing.